+++ packages/ui/src/index.ts
@@ -3,5 +3,8 @@ import "./index.css";
 export { default as FlipCard } from "./components/FlipCard.js";
 export type { FlipCardProps } from "./components/FlipCard.js";
 
+export { BuoyDock } from "./BuoyDock/BuoyDock.js";
+export type { BuoyDockProps } from "./BuoyDock/BuoyDock.js";
+
 export { default as ProactivitySwitch } from "./components/ProactivitySwitch.js";
 export * from "./components/ProactivitySwitch.js";